Add examples readme to ergo-core Move appropriate specs to core
